requirements.txt
asgiref==3.7.2
async-generator==1.10
attrs==22.1.0
autopep8==2.0.0
Brotli==1.1.0
certifi==2022.9.24
cffi==1.15.1
click==8.1.3
colorama==0.4.6
dj-database-url==2.1.0
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.0.0rc9
gunicorn==21.2.0
h11==0.14.0
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
nltk==3.7
outcome==1.2.0
packaging==23.2
pdfkit==1.0.0
plumbum==1.8.1
psycopg2-binary==2.9.9
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.8.0
pypng==0.20220715.0
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PySide2==5.15.2.1
PySocks==1.7.1
pytube==15.0.0
pytz==2023.3
pywin32==306
qrcode==7.4.2
regex==2022.10.31
rpyc==5.3.1
selenium==4.5.0
shiboken2==5.15.2.1
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
tinydb==4.7.1
tomli==2.0.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.12
whitenoise==6.6.0
wsproto==1.2.0
